move another extension class out into its own file
